get progress updates function

+++ FrameMarker/homepage/models.py
@@ -9,7 +9,7 @@ class Video(models.Model):
     title = models.CharField(max_length=255)
     upload_time = models.DateTimeField(auto_now_add=True)
     uploader = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
-    annotation_progress = models.IntegerField(default=0)
+    annotation_progress = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     annotation_time = models.DateTimeField(null=True, blank=True)
     approved = models.BooleanField(default=False)
     video_file = models.FileField(upload_to='Video/')
